+++ rosariosis_mysql.sql
@@ -2103,8 +2103,7 @@ CREATE VIEW transcript_grades AS
             AND sms.student_id = srcg.student_id
         LEFT OUTER JOIN schools
         ON mp.school_id = schools.id
-            AND (mp.mp_source<>'History' AND mp.syear = schools.syear)
-                OR (mp.mp_source='History' AND mp.syear=(SELECT syear FROM schools WHERE mp.school_id = id ORDER BY syear LIMIT 1))
+            AND mp.syear = schools.syear
     ORDER BY srcg.course_period_id;